requirements.txt
allure-pytest                2.9.45
allure-python-commons        2.9.45
pytest                       7.1.2
pytest_check                 1.0.5
pytest-metadata              2.0.1
requests                     2.26.0